backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
aiosqlite
python-jose[cryptography]
passlib
python-multipart
python-dotenv
pydantic
pydantic-settings
websocket
redis
celery
pillow
aiofiles
httpx
email-validator
jinja2
starlette
python-socketio
websockets
